Move runtime input protocol types
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
